# api/docs/swagger.yaml
basePath: /localeats.uz
definitions:
...
  models.Error:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  order.DishStats:
    properties:
      id:
        type: string
      name:
        type: string
      orders_count:
        type: integer
      revenue:
        type: number
    type: object
  order.HourStats:
    properties:
      hour:
        type: string
      orders_count:
        type: integer
      revenue:
        type: number
    type: object
  order.Item:
    properties:
      dish_id:
        type: string
      quantity:
        type: integer
    type: object
  order.KitchenStatistics:
    properties:
      average_rating:
        type: number
      busiest_hours:
        items:
          $ref: '#/definitions/order.HourStats'
        type: array
      top_dishes:
        items:
          $ref: '#/definitions/order.DishStats'
        type: array
      total_orders:
        type: integer
      total_revenue:
        type: number
    type: object
  order.OrderInfo:
    properties:
      created_at:
        type: string
      delivery_address:
        type: string
      delivery_time:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/order.Item'
        type: array
      kitchen_id:
        type: string
      status:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  order.ReqCreateOrder:
    properties:
      delivery_address:
        type: string
      delivery_time:
        type: string
      items:
        items:
          $ref: '#/definitions/order.Item'
        type: array
      kitchen_id:
        type: string
      user_id:
        type: string
    type: object
  order.Status:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
  order.StatusRes:
    properties:
      id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  payment.PaymentInfo:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      order_id:
        type: string
      status:
        type: string
      transaction_id:
        type: string
      updated_at:
        type: string
    type: object
  payment.ReqCreatePayment:
    properties:
      card_number:
        type: string
      cvv:
        type: string
      expiry_date:
        type: string
      order_id:
        type: string
      payment_method:
        type: string
    type: object
...
host: localhost:8888
info:
  contact: {}
  description: LocalEats is a program to order local and homemade food with quality
    and precise delivery.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: LocalEats API Gateway
  version: "1.0"
paths:
...
  /orders/{id}:
    get:
      consumes:
      - application/json
      description: gets
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order informations
          schema:
            $ref: '#/definitions/order.OrderInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets order
      tags:
      - Orders
  /orders/{id}/update:
    put:
      consumes:
      - application/json
      description: updates
      parameters:
      - description: Status information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.Status'
      produces:
      - application/json
      responses:
        "200":
          description: Order informations
          schema:
            $ref: '#/definitions/order.StatusRes'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: updates order
      tags:
      - Orders
  /orders/{kitchen_id}:
    delete:
      consumes:
      - application/json
      description: deletes order
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: deletes order
      tags:
      - Orders
  /orders/chef/{id}:
    get:
      consumes:
      - application/json
      description: gets
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order informations
          schema:
            $ref: '#/definitions/order.OrderInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets order
      tags:
      - Orders
  /orders/create:
    post:
      consumes:
      - application/json
      description: creates
      parameters:
      - description: Order information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/order.ReqCreateOrder'
      produces:
      - application/json
      responses:
        "201":
          description: Order informations
          schema:
            $ref: '#/definitions/order.OrderInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Creates order
      tags:
      - Orders
  /orders/favourites/{user_id}:
    get:
      consumes:
      - application/json
      description: Get statistics
      parameters:
      - description: user id
        in: path
        name: user_id
        required: true
        type: string
      - description: Start date in YYYY-MM-DD format
        in: query
        name: start_date
        required: true
        type: string
      - description: End date in YYYY-MM-DD format
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of kitchens
          schema:
            $ref: '#/definitions/order.KitchenStatistics'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: No Auth
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets statistics
      tags:
      - Orders
  /orders/statistics/{kitchen_id}:
    get:
      consumes:
      - application/json
      description: Get statistics of all available kitchens within a specified date
        range
      parameters:
      - description: kitchen id
        in: path
        name: kitchen_id
        required: true
        type: string
      - description: Start date in YYYY-MM-DD format
        in: query
        name: start_date
        required: true
        type: string
      - description: End date in YYYY-MM-DD format
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of kitchens
          schema:
            $ref: '#/definitions/order.KitchenStatistics'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: No Auth
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets statistics
      tags:
      - Orders
  /orders/user/{id}:
    get:
      consumes:
      - application/json
      description: gets
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        required: true
        type: integer
      - description: page
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order informations
          schema:
            $ref: '#/definitions/order.OrderInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: gets order
      tags:
      - Orders
  /payments/create:
    post:
      consumes:
      - application/json
      description: creates
      parameters:
      - description: Payment information
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/payment.ReqCreatePayment'
      produces:
      - application/json
      responses:
        "201":
          description: Payment informations
          schema:
            $ref: '#/definitions/payment.PaymentInfo'
        "400":
          description: 'Invalid inputs can result to '
          schema:
            $ref: '#/definitions/models.Error'
        "401":
          description: 'No Auth thats the problem '
          schema:
            $ref: '#/definitions/models.Error'
        "500":
          description: Something went wrong in server
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - ApiKeyAuth: []
      summary: Creates payment
      tags:
      - Payments
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"